app/requirements.txt
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.0
filelock==3.8.0
h11==0.14.0
huggingface-hub==0.10.0
idna==3.4
importlib-metadata==5.0.0
numpy==1.21.6
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
pydantic==1.10.2
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.4
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.1
transformers==4.22.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
zipp==3.8.1
python-jose==3.3.0
